Refactor script commands in composer.json
This commit refactors the script commands in the composer.json file. Some commands have been replaced and others changed in order to improve the efficiency of the testing and validation process. More specifically, testing commands have been enhanced for more effective execution with Pest. Also, the styling of the code is now handled with Pint, ensuring that the code adheres to the styling rules without manual involvement by automatically committing significant styling changes.