Improve project structure and upgrade testing level
Deleted the .php-cs-fixer.php and phpunit.xml files to simplify the project structure and avoid duplication, also renamed and updated phpstan and infection configuration files to increase the PHPStan analysis level from 6 to 7 and switch the logs location. The .gitignore was updated to ignore the build/ directory and a new phpunit.xml.dist file was created to improve the test configuration. The Phony.php file was updated adding a TODO comment to remind to implement a mixin analogy to the Laravel IDE helper. These changes were made to make the project more maintainable and with higher code quality standards.