Refactor CI/CD workflows to use phonyland configuration
Removed custom configurations of the GitHub actions workflows for better maintainability. We replaced the "test" job with the "phony_workflows" job that uses a predefined workflow from the phonyland development tools. This change makes it significantly easier to update and maintain compliance with best practices for software testing. It also increases the PHPStan analysis level for a more detailed review of potential code issues.